requirements.txt
# متطلبات تشغيل بوت تيليجرام
python-telegram-bot==20.8
python-dotenv==1.0.1
